+++ src/app/docs/animations/hovers/page.mdx
@@ -1,5 +1,5 @@
 import { RequiresTailwind } from '@/components/Requires'
-import HoverAnimations from '@/components/cards/HoverAnimationCards'
+import HoverAnimations from '@/components/cards_temp/HoverAnimationCards'
 
 # Hover Animations
