+++ src/app/docs/components/button/page.mdx
@@ -1,4 +1,4 @@
-import ButtonCards from '@/components/cards/ButtonCards'
+import ButtonCards from '@/components/cards_temp/ButtonCards'
 
 export const metadata = {
   title: 'Button',
