+++ src/app/docs/components/toggle/page.mdx
@@ -1,4 +1,4 @@
-import ToggleCards from '@/components/cards/ToggleCards'
+import ToggleCards from '@/components/cards_temp/ToggleCards'
 
 export const metadata = {
   title: 'Toggle',
