+++ src/components/cards_temp/AnimationCards.tsx
